# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000"
info:
  title: Fantastic Coffee (decaffeinated)
  description: Fantastic Coffee (decaffeinated), the skeleton project for Web and Software architecture course
  version: 1.0.0
paths:
...
  /name:
    get:
      tags: ["Name"]
      operationId: getName
      summary: Read the stored name
      description: |-
        Returns the name stored in the example table. If no name has been set
        yet, the server replies with 404.
      responses:
        "200":
          description: Name found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NameInfo"
        "404":
          description: No name has been set yet
        "500": { $ref: "#/components/responses/InternalServerError" }
    put:
      tags: ["Name"]
      operationId: setName
      summary: Store a name
      description: |-
        Stores (or replaces) the name in the example table. The stored name is
        returned in the response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameInfo"
      responses:
        "200":
          description: Name stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NameInfo"
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  schemas:
...
    NameInfo:
      title: Name Info
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Jerry
  responses:
...
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
  securitySchemes:
    LegacyAuth:
      type: apiKey
      in: header
      name: X-App-Token
      description: |-
        App session token value. This value is sent by the server in the login
        and session refresh APIs.
...